// app/pages/subscription/interfaces/IFirstStep.ts
export interface IFirstStep {
    cnpj: number;
    nomeFantasia: string;
    cliente: string;
    nomeContato: string;
    telefone: number;
    email: string;
    cpf: number;
    logradouro: string;
    numero: number;
    complemento: string;
    bairro: string;
    pais: string;
    uf: string;
    cidade: string;
    nome: string;
    cpfRepresentante: number;
    emailRepresentante: string;
  }
  